refactor: merged code into program-progress directory
