--- src/components/program-listing/styles/index.scss
@@ -1,1 +1,0 @@
-@import "ProgramListingCard";
