+++ src/components/program-progress/data/hooks.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { logError } from '@edx/frontend-platform/logging';
 import { camelCaseObject } from '@edx/frontend-platform/utils';
-import { getLearnerProgramProgressDetail } from './service';
+import { getLearnerProgramProgressDetail, getLearnerProgramsList } from './service';
 
 export function useLearnerProgramProgressData(programUUID) {
   const [learnerProgramProgressData, setLearnerProgramProgressData] = useState();
@@ -23,4 +23,26 @@ export function useLearnerProgramProgressData(programUUID) {
     fetchData();
   }, [programUUID]);
   return [camelCaseObject(learnerProgramProgressData), fetchError];
+}
+
+export function useLearnerProgramsListData(enterpriseUuid) {
+  const [learnerProgramsListData, setLearnerProgramsListData] = useState();
+  const [fetchError, setFetchError] = useState();
+
+  useEffect(() => {
+    const fetchData = async () => {
+      if (enterpriseUuid) {
+        try {
+          const { data } = await getLearnerProgramsList(enterpriseUuid);
+          setLearnerProgramsListData(data);
+        } catch (error) {
+          logError(error);
+          setFetchError(error);
+        }
+      }
+      return undefined;
+    };
+    fetchData();
+  }, [enterpriseUuid]);
+  return [camelCaseObject([]), fetchError];
 }
