+++ src/components/program-progress/data/service.js
@@ -6,3 +6,9 @@ export function getLearnerProgramProgressDetail(programUUID) {
   const url = `${config.LMS_BASE_URL}/api/dashboard/v0/programs/${programUUID}/progress_details/`;
   return getAuthenticatedHttpClient().get(url);
 }
+
+export function getLearnerProgramsList(enterpriseUuid) {
+  const config = getConfig();
+  const url = `${config.LMS_BASE_URL}/api/dashboard/v0/programs/${enterpriseUuid}/`;
+  return getAuthenticatedHttpClient().get(url);
+}
