+++ src/components/program-progress/styles/_ProgramListingCard.scss
@@ -1,6 +1,10 @@
 .program-listing-card{
   width: 100%;
   max-width: 500px;
+  cursor: pointer;
+}
+.program-listing-card:hover{
+  box-shadow: 0 0.0625rem 0.125rem rgb(0 0 0 / 20%) !important;
 }
 
 .program-progress{
@@ -15,7 +19,6 @@
       border-radius: 50%;
       height: 25px;
       width: 25px;
-      background: green;
       text-align: center;
       line-height: 1.3;
       color: white;
@@ -28,11 +31,11 @@
 }
 
 .completed-courses{
-  background: #04262B!important;
+  background: #0D7D4D !important;
 }
 
 .in-progress-courses{
-  background: #04262B!important;
+  background: #04262B !important;
 }
 
 .no-programs-message{
