+++ src/components/program-progress/styles/index.scss
@@ -1,0 +1,1 @@
+@import "src/components/program-progress/styles/ProgramListingCard";
